+++ CHANGES.md
@@ -2,6 +2,8 @@
 -----------
 
 * Add `:ignore_case` to Slop options for case insensitive option matching
+* Add `:on_noopts` for triggering an event when the arguments contain no
+  options
 
 1.5.5 (2011-05-03)
 ------------------